fix: rename portal.svelte.ts to portal-state.svelte.ts to fix missing generated types
Since the compilation of `portal.svelte` outputs inferred type definitions to the same file as the compilation of `portal.svelte.ts`, one of the files must be renamed to avoid conflicts. This change renames `portal.svelte.ts` to `portal-state.svelte.ts`, which is more descriptive and avoids the conflict.
