+++ src/lib/index.ts
@@ -1,1 +1,1 @@
-export { createPortal } from "./portal.svelte.js";
+export { createPortal } from "./portal-state.svelte.js";
