+++ src/lib/portal.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { untrack } from "svelte";
-    import { type ChildrenSnippet, PortalState } from "./portal.svelte.js";
+    import { type ChildrenSnippet, PortalState } from "./portal-state.svelte.js";
     import { BROWSER } from "esm-env";
 
     type Props = {
